src/test/java/aldora/spring/recipe/services/ImageServiceImplTest.java
package aldora.spring.recipe.services;

import aldora.spring.recipe.model.Recipe;
import aldora.spring.recipe.repositories.reactive.RecipeReactiveRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.mock.web.MockMultipartFile;
import org.springframework.web.multipart.MultipartFile;
import reactor.core.publisher.Mono;

import java.io.IOException;
import java.util.Optional;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.*;

public class ImageServiceImplTest {
    @Mock
    RecipeReactiveRepository recipeReactiveRepository;

    ImageService imageService;

    @Captor
    ArgumentCaptor<Recipe> recipeArgumentCaptor;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        imageService = new ImageServiceImpl(recipeReactiveRepository);
    }

    @Test
    public void saveImageFile() throws IOException {
        String id = "1";

        MultipartFile image = new MockMultipartFile("imagefile", "test.txt", "text/plain",
                "sping framework guru".getBytes());

        Recipe recipe = new Recipe();
        recipe.setId(id);

        when(recipeReactiveRepository.findById(anyString())).thenReturn(Mono.just(recipe));
        when(recipeReactiveRepository.save(any())).thenReturn(Mono.just(recipe));

        imageService.saveImageFile(anyString(), image);

        verify(recipeReactiveRepository).save(recipeArgumentCaptor.capture());

        Recipe savedRecipe = recipeArgumentCaptor.getValue();
        assertEquals(id, savedRecipe.getId());
        assertEquals(image.getBytes().length, savedRecipe.getImage().length);
    }
}